+++ src/main/java/blusunrize/immersiveengineering/common/util/FakePlayerUtil.java
@@ -9,22 +9,43 @@
 package blusunrize.immersiveengineering.common.util;
 
 import com.mojang.authlib.GameProfile;
+import net.minecraft.world.World;
 import net.minecraft.world.WorldServer;
 import net.minecraftforge.common.util.FakePlayer;
 import net.minecraftforge.common.util.FakePlayerFactory;
+import net.minecraftforge.event.world.WorldEvent;
+import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
+import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
+import java.util.Map;
 import java.util.UUID;
+import java.util.WeakHashMap;
 
+@EventBusSubscriber
 public class FakePlayerUtil
 {
 	private static GameProfile IE_PROFILE = new GameProfile(UUID.fromString("99562b85-bd1a-4ded-bb1a-c307bf0c0133"), "[ImmersiveEngineering]");
-	private static FakePlayer fakePlayerInstance;
+	private static Map<World, FakePlayer> fakePlayerInstances = new WeakHashMap<>();
 
-	public static FakePlayer getFakePlayer(WorldServer world)
+	public static FakePlayer getAnyFakePlayer()
 	{
-		if(fakePlayerInstance==null)
-			fakePlayerInstance = FakePlayerFactory.get(world, IE_PROFILE);
-		return fakePlayerInstance;
+		if(fakePlayerInstances.isEmpty())
+			return null;
+		else
+			return fakePlayerInstances.values().iterator().next();
+	}
+
+	public static FakePlayer getFakePlayer(World w)
+	{
+		return fakePlayerInstances.get(w);
+	}
+
+	@SubscribeEvent
+	public static void onLoad(WorldEvent.Load ev)
+	{
+		World world = ev.getWorld();
+		if(world instanceof WorldServer)
+			fakePlayerInstances.put(world, FakePlayerFactory.get((WorldServer)world, IE_PROFILE));
 	}
 
 }